client/requirements.txt
opencv-python
numpy
pyzmq
adafruit-circuitpython-servokit
argparse